add exception Handler, fix some bugs, add init for Database

+++ src/main/java/com/goncharov/tinyurl/controller/RedirectController.java
@@ -1,9 +1,9 @@
 package com.goncharov.tinyurl.controller;
 
 import com.goncharov.tinyurl.dto.UrlDto;
-import com.goncharov.tinyurl.dto.UrlErrorDto;
-import com.goncharov.tinyurl.dto.UrlResponseDto;
+import com.goncharov.tinyurl.dto.UrlRequestDto;
 import com.goncharov.tinyurl.entity.Url;
+import com.goncharov.tinyurl.exception.ApiRequestException;
 import com.goncharov.tinyurl.mapper.UrlMapper;
 import com.goncharov.tinyurl.service.UrlService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -24,51 +24,60 @@ public class RedirectController {
         this.urlService = urlService;
     }
 
-    @PostMapping("/generate")
-    public ResponseEntity<?> generateShortLink(@RequestBody UrlDto urlDto) {
-        Url url = urlService.createUrl(urlDto);
-
-        if (url != null) {
-            return new ResponseEntity<>(UrlMapper.INSTANCE.toDto(url), HttpStatus.OK);
+    @PostMapping(value = "/generate")
+    public ResponseEntity<?> generateShortLink(@RequestBody(required = false) UrlDto urlDto) throws ApiRequestException {
+        if (urlDto == null) {
+            throw new ApiRequestException("Url must be not null.");
         }
 
-        UrlErrorDto errorDto = new UrlErrorDto();
-        errorDto.setStatus("404");
-        errorDto.setError("Oops! Unexpected error, try again later!");
+        if (urlDto.getUrl().length() < 4) {
+            throw new ApiRequestException("Oops! Url is to short.");
+        }
 
-        return new ResponseEntity<>(errorDto, HttpStatus.BAD_GATEWAY);
+        Url url = urlService.createUrl(urlDto);
+
+        return new ResponseEntity<>(UrlMapper.INSTANCE.toDto(url), HttpStatus.OK);
     }
 
-    @GetMapping("/{alias}")
+    @GetMapping(value = "/{alias}")
     public ResponseEntity<?> redirectToOriginalUrl
-            (@PathVariable String alias, HttpServletResponse response) throws IOException {
+            (@PathVariable String alias, HttpServletResponse response) throws IOException, ApiRequestException {
         Url url = urlService.getUrlFromAlias(alias);
 
         if (url == null) {
-            UrlErrorDto errorDto = new UrlErrorDto();
-            errorDto.setError("Url doesn't exist or has expired");
-            errorDto.setStatus("400");
-
-            return new ResponseEntity<>(errorDto, HttpStatus.BAD_REQUEST);
+            throw new ApiRequestException("Url is outdated or was never created.");
         }
 
-        urlService.incrementCounter(url);
         response.sendRedirect(url.getUrl());
+        urlService.incrementCounter(url);
 
         return new ResponseEntity<>(HttpStatus.MOVED_PERMANENTLY);
     }
 
     @PostMapping("/delete")
-    public ResponseEntity<?> deleteAlias(@RequestBody UrlResponseDto response) {
-        Url url = urlService.getUrlFromAlias(response.getAlias());
+    public ResponseEntity<?> deleteAlias(@RequestBody(required = false) UrlRequestDto requestDto) throws ApiRequestException {
+        if (requestDto == null) {
+            throw new ApiRequestException("Url must be not null.");
+        }
+
+        Url url = urlService.getUrlFromAlias(requestDto.getAlias());
+
+        if (url == null) {
+            throw new ApiRequestException("Url is outdated or was never created.");
+        }
+
         urlService.deleteUrl(url);
 
         return new ResponseEntity<>(HttpStatus.OK);
     }
 
     @GetMapping("/info/{alias}")
-    public ResponseEntity<?> getInfo(@PathVariable String alias) {
+    public ResponseEntity<?> getInfo(@PathVariable String alias) throws ApiRequestException {
         Url url = urlService.getUrlFromAlias(alias);
+
+        if (url == null) {
+            throw new ApiRequestException("Url is outdated or was never created.");
+        }
 
         return new ResponseEntity<>(UrlMapper.INSTANCE.toDto(url), HttpStatus.OK);
     }
